requirements.txt
fastapi
uvicorn
pydantic
newspaper3k
beautifulsoup4
lxml_html_clean
